first_approach_llm/requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
pydantic>=1.9.0
pydantic-ai>=0.1.0
logfire
google.generativeai
bs4